requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.9.2
langchain==0.3.1
langchain-community==0.3.1
langchain-openai==0.2.1
langgraph==0.2.32
faiss-cpu==1.9.0.post1
requests==2.32.3
jinja2==3.1.4
openai==1.50.2
httpx==0.27.2
websockets==12.0
pypdf==4.3.1
python-docx==1.1.2
openpyxl==3.1.5
xlrd==2.0.1
